modified

+++ Admin/ajax/features_facilities.php
@@ -43,11 +43,17 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_feature']];
 
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q,$values,'i');
-        echo $res;
-    }
+        $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?',[$frm_data['rem_feature']],'i');
 
+        if(mysqli_num_rows($check_q)==0)
+        {
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q,$values,'i');
+            echo $res;
+        } else {
+            echo 'room_added';
+        }       
+    }
 
     if(isset($_POST['add_facility']))
     {
@@ -68,7 +74,6 @@
         }
 
     }
-
 
     if(isset($_POST['get_facilities']))
     {
@@ -101,22 +106,26 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_facility']];
 
+        $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?',[$frm_data['rem_facility']],'i');
 
-        $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
-        $res = select($pre_q,$values,'i');
-        $img = mysqli_fetch_assoc($res);
+        if(mysqli_num_rows($check_q)==0)
+        {
+            $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
+            $res = select($pre_q,$values,'i');
+            $img = mysqli_fetch_assoc($res);
 
-        if(deleteImage($img['icon'], FACILITY_FOLDER)){
-            $q = "DELETE FROM `facilities` WHERE `id`=?";
-            $res = delete($q,$values,'i');
-            echo $res;
+            if(deleteImage($img['icon'], FACILITY_FOLDER)){
+                $q = "DELETE FROM `facilities` WHERE `id`=?";
+                $res = delete($q,$values,'i');
+                echo $res;
+            } 
+            else {
+                echo 0;
+            }  
         } 
         else {
-            echo 0;
-        }
-
-        
+            echo 'room_added';
+        }               
     }
-
 
 ?>
